+++ kafka-rest/client/src/main/kotlin/org/http4k/connect/kafka/rest/action/ConsumeRecords.kt
@@ -3,7 +3,7 @@ package org.http4k.connect.kafka.rest.action
 import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.kClass
 import org.http4k.connect.kafka.rest.model.ConsumerGroup
-import org.http4k.connect.kafka.rest.model.ConsumerInstanceId
+import org.http4k.connect.kafka.rest.model.ConsumerInstance
 import org.http4k.connect.kafka.rest.model.RecordFormat
 import org.http4k.connect.kafka.rest.model.TopicRecord
 import org.http4k.core.Method.GET
@@ -13,7 +13,7 @@ import java.time.Duration
 @Http4kConnectAction
 data class ConsumeRecords(
     val group: ConsumerGroup,
-    val instance: ConsumerInstanceId,
+    val instance: ConsumerInstance,
     val format: RecordFormat,
     val timeout: Duration? = null
 ) :