+++ kafka-rest/client/src/main/kotlin/org/http4k/connect/kafka/rest/action/SeekOffsets.kt
@@ -4,7 +4,7 @@ import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.kClass
 import org.http4k.connect.kafka.rest.KafkaRestMoshi.auto
 import org.http4k.connect.kafka.rest.model.ConsumerGroup
-import org.http4k.connect.kafka.rest.model.ConsumerInstanceId
+import org.http4k.connect.kafka.rest.model.ConsumerInstance
 import org.http4k.connect.kafka.rest.model.SeekOffset
 import org.http4k.connect.kafka.rest.model.SeekOffsetsSet
 import org.http4k.core.Body
@@ -17,7 +17,7 @@ import org.http4k.core.with
 @Http4kConnectAction
 data class SeekOffsets(
     val group: ConsumerGroup,
-    val instance: ConsumerInstanceId,
+    val instance: ConsumerInstance,
     val offsets: List<SeekOffset>
 ) : KafkaRestAction<Unit>(kClass()) {
     override fun toRequest() = Request(POST, "/consumers/$group/instances/$instance/positions")