+++ kafka-rest/client/src/main/kotlin/org/http4k/connect/kafka/rest/action/consumer/Delete.kt
@@ -7,5 +7,5 @@ import org.http4k.core.Request
 
 @Http4kConnectAction
 object Delete : KafkaRestConsumerAction<Unit>(kClass()) {
-    override fun toRequest() = Request(DELETE, "/consumers/{group}/instances/{instance}")
+    override fun toRequest() = Request(DELETE, "")
 }