+++ kafka-rest/client/src/test/kotlin/org/http4k/connect/kafka/rest/extensions/ProduceMessagesWithPartitionsTest.kt
@@ -1,4 +1,4 @@
-package org.http4k.connect.kafka.rest.action
+package org.http4k.connect.kafka.rest.extensions
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
@@ -6,15 +6,15 @@ import dev.forkhandles.result4k.valueOrNull
 import org.http4k.connect.kafka.rest.FakeKafkaRest
 import org.http4k.connect.kafka.rest.Http
 import org.http4k.connect.kafka.rest.KafkaRest
+import org.http4k.connect.kafka.rest.action.ProducedMessages
 import org.http4k.connect.kafka.rest.model.Record
 import org.http4k.connect.kafka.rest.model.Records
 import org.http4k.connect.kafka.rest.model.Topic
-import org.http4k.connect.kafka.rest.partitioning.RoundRobinRecordPartitioner
 import org.http4k.core.Credentials
 import org.http4k.core.Uri
 import org.junit.jupiter.api.Test
 
-class ProduceMessagesKtTest {
+class ProduceMessagesWithPartitionsTest {
 
     @Test
     fun `writing to a list of partitions using a partitioner`() {
@@ -23,9 +23,14 @@ class ProduceMessagesKtTest {
         )
 
         assertThat(
-            kafkaRest.produceMessages(Topic.of("asd"), Records.Json(listOf(Record("123", ""))), ::RoundRobinRecordPartitioner)
+            kafkaRest.produceMessagesWithPartitions(
+                Topic.of("topic"),
+                Records.Json(listOf(Record("123", ""))),
+                ::RoundRobinRecordPartitioner
+            )
                 .valueOrNull()!!,
             equalTo(ProducedMessages(null, null, listOf()))
         )
     }
 }
+