+++ kafka-rest/fake/src/main/kotlin/org/http4k/connect/kafka/rest/endpoints/GetOffsets.kt
@@ -24,6 +24,7 @@ fun getOffsets(consumers: Storage<ConsumerState>) =
         val group = Path.value(ConsumerGroup).of("consumerGroup")(req)
         consumers[group]
             ?.let {
+                println(it)
                 Response(OK)
                     .with(Body.auto<CommitOffsetsSet>().toLens() of CommitOffsetsSet(
                         it.offsets