+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v
+- **http4k-connect-kakfa-rest*** - Add helpers for consuming and producing.
+
 ### v3.31.1.0
 - **http4k-connect-kakfa-rest*** - Adding message partitioning strategies.