+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.33.1.0
+- **http4k-connect-kakfa-schemaregistry*** - Changes to register schema version API contract.
+
 ### v3.33.0.0
 - **http4k-connect-*** - Upgrade dependencies,
-- **http4k-connect-kakfa-http-proxy*** - New module.
+- **http4k-connect-kakfa-schemaregistry*** - New module & fakes
 
 ### v3.32.0.0
 - **http4k-connect-kakfa-rest*** - Add helpers for consuming and producing.